requirements.txt
streamlit>=1.33.0
requests>=2.31.0
python-dotenv>=1.0.1
pillow>=10.2.0
numpy>=1.26.0
tensorflow>=2.15.0
